commiting for security

+++ FarmerApi/src/main/java/com/cropdeals/FarmerApi/controller/FarmerController.java
@@ -5,6 +5,9 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,9 +18,10 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
  
 import com.cropdeals.FarmerApi.Services.ProfileServices;
- 
+import com.cropdeals.FarmerApi.exception.ApiRequestException;
 import com.cropdeals.FarmerApi.models.FarmerProfile;
 import com.cropdeals.FarmerApi.models.ReturnAllCrops;
+import com.cropdeals.FarmerApi.models.AuthenticationResponse;
 import com.cropdeals.FarmerApi.models.Crops;
 import com.cropdeals.FarmerApi.repository.ProfileRepo;
  
@@ -37,19 +41,38 @@ public class FarmerController {
 	ProfileRepo profileRepo;
 	
 	
+	@Autowired
+	private AuthenticationManager authenticationManager;
+	
 	@PostMapping("/addfarmer")
-	public void savefarmer( @RequestBody FarmerProfile farmerProfile ) {
-		profileServices.addfarmer( farmerProfile );
+	public String addfarmer( @RequestBody FarmerProfile farmerProfile ) {
+		profileRepo.save( farmerProfile );
+		return "farmer added";
+	}
+	
+	@PostMapping("/auth")
+	private ResponseEntity<?> authfarmer(@RequestBody  FarmerProfile farmerProfile){
+	String Name = farmerProfile.getfName();
+	String Password = farmerProfile.getfPassword ();
+	try {
+	authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(Name, Password));
+
+	} catch (Exception e) {
+
+	return ResponseEntity.ok(new AuthenticationResponse("Error during customer Authentication"+ Name));
+	}
+	return ResponseEntity.ok(new AuthenticationResponse("Successfully Authenticated customer "+ Name));
+
 	}
 	
 	@PostMapping("/addcrop")
 	public void addCrops(@RequestBody Crops crop ) {
-		restTemplate.postForEntity("http://crop-avalilable/crops/addcrop", crop,Crops.class);
-	}
+		restTemplate.postForEntity("http://crop-avalilable/crops/addcrop", crop , Crops.class);
+	} 
 
 	@GetMapping("/allfarmers")
 	public List<FarmerProfile> showAllFarmers() {
-		return profileServices.findAll();		
+		return profileRepo.findAll();		
 	}
 	
 	
@@ -57,21 +80,30 @@ public class FarmerController {
 	@GetMapping("/allmycrops/{farmerId}")
 	public ReturnAllCrops showMyCrops(@PathVariable String farmerId) {
 		
-		return  restTemplate.getForObject("http://localhost:8084/crops/findmycrop/"+farmerId, ReturnAllCrops.class);
+		return  restTemplate.getForObject("http://crop-avalilable/crops/findmycrop/"+farmerId, ReturnAllCrops.class);
 		
 	}
 	
 	@GetMapping("/allcrops")
-	public List<Crops> showAllCrops() {
+	public String showAllCrops() {
 		
-		ReturnAllCrops allCrops = restTemplate.getForObject("http://localhost:8084/crops/allcrops", ReturnAllCrops.class);
-		return allCrops.getListOfCrops();
+		return restTemplate.getForObject("http://crop-avalilable/crops/allcrops", String.class);
+		
 	}
 	
 	@GetMapping("/findfarmer/{farmerName}")
 	public FarmerProfile findfarmer ( @PathVariable String farmerName ){
 		return profileRepo.findfarmerByName( farmerName );
 	}
+	
+	@PutMapping("/rating/{id}")
+    public void giveRating(@RequestBody FarmerProfile farmerProfile, @PathVariable String id) {
+        farmerProfile.setId( id );  
+        Integer rating= farmerProfile.getAllRatings().stream().mapToInt(Integer::intValue).sum()/farmerProfile.getAllRatings().size();
+        farmerProfile.setRating(rating);
+        profileRepo.save(farmerProfile);
+    }
+	
 	
 	@PutMapping("/update/{id}")
     public String updateOrder(@RequestBody FarmerProfile order, @PathVariable String id) {
@@ -83,13 +115,28 @@ public class FarmerController {
 	@GetMapping("/deletecrop/{id}")
 	public String deleteCrop( @PathVariable String id )	{
 		
-		 return restTemplate.getForObject("http://localhost:8084/crops/delete/"+id, String.class);
+		 return restTemplate.getForObject("http://crop-avalilable/crops/delete/"+id, String.class);
 
 	}
 	
+	
+	
+	
+	
 	@GetMapping("/deletefarmer/{id}")
 	public String deleteFarmer( @PathVariable String id )	{
-		profileRepo.deleteById( id );
-		return ("Deleted Successfully");
-	}
+	{
+		boolean isfarmerExist= profileRepo.existsById(id);
+		if(isfarmerExist) {
+			profileRepo.deleteById( id );
+			return ("Deleted Successfully");
+		}
+		else
+		{
+			throw new ApiRequestException("CAN NOT DELETE AS USER NOT FOUND WITH THIS ID ::");
+		}
+
+	}	
+}
+
 }
